open-grant-proposals: Bidirectional FIL <=> WFIL Bridge to Ethereum
Hi there! Enclosed is a grant request submission in hopes of getting to
be a part of Wave 8, to build an open source, bidirectional bridge
between FIL and WFIL on Ethereum! :)

Thanks so much to you all for the opportunity to apply!

Standing by and happy to answer any questions.